requirements.txt
# service_bot/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
aioredis==2.0.1